Applies code review suggestion: Fixes crash when list is empty and try to clear; Renames Movies and Movies Result; Adds verification if device is connected to internet; Adds button to try again; Simplify code, correct indentation and renames imageView object. Removes unused layout parent. Adds JsonParser class to remove logic from Activity.
