+++ app/src/main/java/com/lsaippa/movies/model/MovieResult.java
@@ -1,231 +1,71 @@
 package com.lsaippa.movies.model;
 
-import android.os.Parcel;
-import android.os.Parcelable;
-
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
 
-import java.util.ArrayList;
 import java.util.List;
 
-public class MovieResult implements Parcelable {
 
-    @SerializedName("vote_count")
+@SuppressWarnings("unused")
+public class MovieResult {
+
+
+    @SerializedName("page")
     @Expose
-    private Integer voteCount;
-    @SerializedName("id")
+    private Integer page;
+    @SerializedName("total_results")
     @Expose
-    private Integer id;
-    @SerializedName("video")
+    private Integer totalResults;
+    @SerializedName("total_pages")
     @Expose
-    private Boolean video;
-    @SerializedName("vote_average")
+    private Integer totalPages;
+    @SerializedName("results")
     @Expose
-    private Double voteAverage;
-    @SerializedName("title")
-    @Expose
-    private String title;
-    @SerializedName("popularity")
-    @Expose
-    private Double popularity;
-    @SerializedName("poster_path")
-    @Expose
-    private String posterPath;
-    @SerializedName("original_language")
-    @Expose
-    private String originalLanguage;
-    @SerializedName("original_title")
-    @Expose
-    private String originalTitle;
-    @SerializedName("genre_ids")
-    @Expose
-    private List<Integer> genreIds = null;
-    @SerializedName("adult")
-    @Expose
-    private Boolean adult;
-    @SerializedName("overview")
-    @Expose
-    private String overview;
-    @SerializedName("release_date")
-    @Expose
-    private String releaseDate;
+    private List<Movies> results = null;
 
-    public Integer getVoteCount() {
-        return voteCount;
+    public MovieResult(){
+
     }
 
-    public void setVoteCount(Integer voteCount) {
-        this.voteCount = voteCount;
+    public Integer getPage() {
+        return page;
     }
 
-    public Integer getId() {
-        return id;
+    public void setPage(Integer page) {
+        this.page = page;
     }
 
-    public void setId(Integer id) {
-        this.id = id;
+    public Integer getTotal_results() {
+        return totalResults;
     }
 
-    public Boolean getVideo() {
-        return video;
+    public void setTotal_results(Integer total_results) {
+        this.totalResults = total_results;
     }
 
-    public void setVideo(Boolean video) {
-        this.video = video;
+    public Integer getTotal_pages() {
+        return totalPages;
     }
 
-    public Double getVoteAverage() {
-        return voteAverage;
+    public void setTotal_pages(Integer total_pages) {
+        this.totalPages = total_pages;
     }
 
-    public void setVoteAverage(Double voteAverage) {
-        this.voteAverage = voteAverage;
+    public List<Movies> getResults() {
+        return results;
     }
 
-    public String getTitle() {
-        return title;
-    }
-
-    public void setTitle(String title) {
-        this.title = title;
-    }
-
-    public Double getPopularity() {
-        return popularity;
-    }
-
-    public void setPopularity(Double popularity) {
-        this.popularity = popularity;
-    }
-
-    public String getPosterPath() {
-        return posterPath;
-    }
-
-    public void setPosterPath(String posterPath) {
-        this.posterPath = posterPath;
-    }
-
-    public String getOriginalLanguage() {
-        return originalLanguage;
-    }
-
-    public void setOriginalLanguage(String originalLanguage) {
-        this.originalLanguage = originalLanguage;
-    }
-
-    public String getOriginalTitle() {
-        return originalTitle;
-    }
-
-    public void setOriginalTitle(String originalTitle) {
-        this.originalTitle = originalTitle;
-    }
-
-    public List<Integer> getGenreIds() {
-        return genreIds;
-    }
-
-    public void setGenreIds(List<Integer> genreIds) {
-        this.genreIds = genreIds;
-    }
-
-    public Boolean getAdult() {
-        return adult;
-    }
-
-    public void setAdult(Boolean adult) {
-        this.adult = adult;
-    }
-
-    public String getOverview() {
-        return overview;
-    }
-
-    public void setOverview(String overview) {
-        this.overview = overview;
-    }
-
-    public String getReleaseDate() {
-        return releaseDate;
-    }
-
-    public void setReleaseDate(String releaseDate) {
-        this.releaseDate = releaseDate;
+    public void setResults(List<Movies> results) {
+        this.results = results;
     }
 
     @Override
     public String toString() {
         return "MovieResult{" +
-                "voteCount=" + voteCount +
-                ", id=" + id +
-                ", video=" + video +
-                ", voteAverage=" + voteAverage +
-                ", title='" + title + '\'' +
-                ", popularity=" + popularity +
-                ", posterPath='" + posterPath + '\'' +
-                ", originalLanguage='" + originalLanguage + '\'' +
-                ", originalTitle='" + originalTitle + '\'' +
-                ", genreIds=" + genreIds +
-                ", adult=" + adult +
-                ", overview='" + overview + '\'' +
-                ", releaseDate='" + releaseDate + '\'' +
+                "page=" + page +
+                ", total_results=" + totalResults +
+                ", total_pages=" + totalPages +
+                ", results=" + results +
                 '}';
     }
-
-
-    @Override
-    public int describeContents() {
-        return 0;
-    }
-
-    @Override
-    public void writeToParcel(Parcel dest, int flags) {
-        dest.writeValue(this.voteCount);
-        dest.writeValue(this.id);
-        dest.writeValue(this.video);
-        dest.writeValue(this.voteAverage);
-        dest.writeString(this.title);
-        dest.writeValue(this.popularity);
-        dest.writeString(this.posterPath);
-        dest.writeString(this.originalLanguage);
-        dest.writeString(this.originalTitle);
-        dest.writeList(this.genreIds);
-        dest.writeValue(this.adult);
-        dest.writeString(this.overview);
-        dest.writeString(this.releaseDate);
-    }
-
-    public MovieResult() {
-    }
-
-    protected MovieResult(Parcel in) {
-        this.voteCount = (Integer) in.readValue(Integer.class.getClassLoader());
-        this.id = (Integer) in.readValue(Integer.class.getClassLoader());
-        this.video = (Boolean) in.readValue(Boolean.class.getClassLoader());
-        this.voteAverage = (Double) in.readValue(Double.class.getClassLoader());
-        this.title = in.readString();
-        this.popularity = (Double) in.readValue(Double.class.getClassLoader());
-        this.posterPath = in.readString();
-        this.originalLanguage = in.readString();
-        this.originalTitle = in.readString();
-        this.genreIds = new ArrayList<>();
-        in.readList(this.genreIds, Integer.class.getClassLoader());
-        this.adult = (Boolean) in.readValue(Boolean.class.getClassLoader());
-        this.overview = in.readString();
-        this.releaseDate = in.readString();
-    }
-
-    public static final Creator<MovieResult> CREATOR = new Creator<MovieResult>() {
-        @Override
-        public MovieResult createFromParcel(Parcel source) {
-            return new MovieResult(source);
-        }
-
-        @Override
-        public MovieResult[] newArray(int size) {
-            return new MovieResult[size];
-        }
-    };
 }
