+++ app/src/main/res/layout/movie_list_item.xml
@@ -1,15 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
+<android.support.v7.widget.CardView
     xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    android:orientation="vertical">
-
-    <android.support.v7.widget.CardView
-        android:id="@+id/card_view"
-        android:layout_gravity="center"
-        android:layout_width="@dimen/iv_poster_list_width"
-        android:layout_height="@dimen/iv_poster_list_height">
+    android:id="@+id/card_view"
+    android:layout_gravity="center"
+    android:layout_width="@dimen/iv_poster_list_width"
+    android:layout_height="@dimen/iv_poster_list_height">
 
     <ImageView
         android:id="@+id/iv_movie"
@@ -17,7 +12,6 @@
         android:layout_height="match_parent"
         android:contentDescription="@string/poster_description"/>
 
+</android.support.v7.widget.CardView>
 
-    </android.support.v7.widget.CardView>
 
-</LinearLayout>
