+++ app/src/main/res/values/strings.xml
@@ -7,5 +7,7 @@
     <string name="most_popular">Most popular</string>
     <string name="most_rated">Most rated</string>
     <string name="internet_error">Cannot connect to Internet. Please check your connection!</string>
-    <string name="poster_description">Poster movie image.</string>
+    <string name="poster_description">Poster movies image.</string>
+    <string name="generic_error">Something is wrong!</string>
+    <string name="try_again">Try again</string>
 </resources>
